Move default SSH iptables INPUT rule to template. Add two vars to set SSH rule dport and source address.
Because role default vars are at the bottom of the role precedence chain, the default SSH rule must be included everytime the iptables_filter_rules is defined elsewhere. This exposes the potential the SSH inbound rule will be accidentally left out, preventing Anisble from being able to manage the host in the future. Moving the rule into the template will ensure it is always included, regardless of whether the iptables_filter_rules is defined elsewhere.